Add ItemVariationPicker component
This component renders a list of radio buttons used to display and select from a list of 'item variants'. Currently the only meaningful item variant in WPCOM is the plan length selector, but this component is designed to allow arbitrary kinds of variants. The `WPCOMProductVariant` type represents one of a list of variations on a given product; the label is the text representing the variant and the description component is for displaying the price, discount graphics, etc.

Note the `getItemVariants` prop: this is a function from product slugs to product variants, and what allows this component to be generic to the product type. We'll have to construct this function and pass it in here. This requires touching a lot of global state, so we handle it at the CompositeCheckout component.